修改网络请求接口&任务列表显示(部分)&作业item点击&部分作业详情页&个人信息界面

+++ app/src/main/java/com/mini/homeworks/MyAssign/Bean/Delete.java
@@ -7,8 +7,8 @@ public class Delete extends LitePalSupport {
     private String siteId;
     private int status;
     private String assignName;
-    private int beginTime;
-    private int endTime;
+    private long beginTime;
+    private long endTime;
     private int color;
     private String assignId;
     public  void setColor(int color) {
@@ -27,11 +27,11 @@ public class Delete extends LitePalSupport {
         this.siteId = siteId;
     }
 
-    public void setEndTime(int endTime) {
+    public void setEndTime(long endTime) {
         this.endTime = endTime;
     }
 
-    public void setBeginTime(int beginTime) {
+    public void setBeginTime(long beginTime) {
         this.beginTime = beginTime;
     }
 
@@ -51,11 +51,11 @@ public class Delete extends LitePalSupport {
         return siteId;
     }
 
-    public int getEndTime() {
+    public long getEndTime() {
         return endTime;
     }
 
-    public int getBeginTime() {
+    public long getBeginTime() {
         return beginTime;
     }
